Apply review: move thumbnail loading out of Player
