+++ app/src/main/java/org/schabi/newpipe/fragments/detail/VideoDetailFragment.java
@@ -690,6 +690,7 @@ public final class VideoDetailFragment
                 .into(binding.detailThumbnailImageView, new Callback() {
                     @Override
                     public void onSuccess() {
+                        // nothing to do, the image was loaded correctly into the thumbnail
                     }
 
                     @Override
